start, stop, setInterval and getConfig methods
Added start method to launch or relaunch the instance's interval.

Added stop method to suspend the instance's interval.

Added setInterval method to change the instance's interval.

Added getOptions method to get the instance's options.
Can be used to clone the instance.

The instances' start() method must now be called for them to speak.